trollapp/requirements.txt
Django==2.0.6
django-cors-headers==2.3.0
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
mysqlclient==1.3.12
Pillow==5.1.0
pytz==2018.4
